[#14939] YSQL: Allow admin to limit the number of successive failed login attempts by a role
Summary:
A security requirement is to track the number of failed login attempts and optionally lock the account for a specified period of time.
The feature is gated by flag `ysql_enable_profile` which is false by default.

The feature consists of two parts:
* Add two system catalog tables to track failed attempts and if the role is enabled/disabled
* Augment the auth module to read and update the new system catalog tables.

System Catalog Tables are:

```
\d pg_yb_profile
                  Table "pg_catalog.pg_yb_profile"
         Column         |   Type   | Collation | Nullable | Default
------------------------+----------+-----------+----------+---------
 prfname                | name     |           | not null |
 prffailedloginattempts | smallint |           | not null |
 prfpasswordlocktime    | bigint   |           | not null |
Indexes:
    "pg_yb_profile_oid_index" PRIMARY KEY, lsm (oid ASC)
```

```
\d pg_yb_role_profile
               Table "pg_catalog.pg_yb_role_profile"
         Column         |   Type   | Collation | Nullable | Default
------------------------+----------+-----------+----------+---------
 rolisenabled           | boolean  |           | not null |
 rolid                  | oid      |           | not null |
 prfid                  | oid      |           | not null |
 rolfailedloginattempts | smallint |           | not null |
 rollockedat            | bigint   |           | not null |
Indexes:
    "pg_yb_role_profile_oid_index" PRIMARY KEY, lsm (rolid ASC), tablespace "pg_global"
Tablespace: "pg_global"
```

DDL Statements are provided to CREATE and DROP profiles. There is a default profile which does not limit the number of failed attempts.

The auth module tracks the number of failed attempts in `pg_yb_role_profile.rolfailedloginattempts`.
If the failed attempts exceeds `pg_yb_profile.prffailedloginattempts` then role is DISABLED. Disabled roles are not allowed to login.

The admin account can control the state of the role using ALTER ROLE ... PROFILE statements such as:
- ATTACH or DETACH a role from a profile.
- ENABLE or DISABLE a role.

Test Plan:
TestYSQLUpgrade
TestPgRegressProfile

Reviewers: smishra, skumar, telgersma

Differential Revision: https://phabricator.dev.yugabyte.com/D21043