[#14939] YSQL: Allow admin to limit the number of successive failed login attempts by a role
Summary:
A security requirement is to track the number of failed login attempts and optionally lock the account for a specified period of time.
The feature is gated by a tserver gflag `ysql_enable_profile` which is false by default. The gflag controls if DDL statements for PROFILE are allowed
AND whether authentication code checks the profile system catalog tables. Since the gflag is read only during authentication and DDL execution,
it is OK if there is inconsistency of its value during upgrades.

The feature consists of two parts:
* Add two system catalog tables.
* Augment the auth module to read and update the new system catalog tables.

System Catalog Tables are:

pg_yb_profile stores the parameters of a profile.

```
\d pg_yb_profile
                  Table "pg_catalog.pg_yb_profile"
         Column         |   Type   | Collation | Nullable | Default
------------------------+----------+-----------+----------+---------
 prfname                | name     |           | not null |
 prffailedloginattempts | smallint |           | not null |
 prfpasswordlocktime    | bigint   |           | not null |
Indexes:
    "pg_yb_profile_oid_index" PRIMARY KEY, lsm (oid ASC)
```

pg_yb_role_profile stores the mapping of a role to a profile. It also stores the state of profile parameters for that role.
```
\d pg_yb_role_profile
               Table "pg_catalog.pg_yb_role_profile"
         Column         |   Type   | Collation | Nullable | Default
------------------------+----------+-----------+----------+---------
 rolisenabled           | boolean  |           | not null |
 rolid                  | oid      |           | not null |
 prfid                  | oid      |           | not null |
 rolfailedloginattempts | smallint |           | not null |
 rollockedat            | bigint   |           | not null |
Indexes:
    "pg_yb_role_profile_oid_index" PRIMARY KEY, lsm (rolid ASC), tablespace "pg_global"
Tablespace: "pg_global"
```

DDL Statements are provided to CREATE and DROP profiles. There is a default profile which does not limit the number of failed attempts.

The auth module tracks the number of failed attempts in `pg_yb_role_profile.rolfailedloginattempts`.
If the failed attempts exceeds `pg_yb_profile.prffailedloginattempts` then role is DISABLED. Disabled roles are not allowed to login.

The admin account can control the state of the role using ALTER ROLE ... PROFILE statements such as:
- ATTACH or DETACH a role from a profile.
- ENABLE or DISABLE a role.

TODO:
[x] Fix Initdb/Reinitdb and make TestYSQLUpgrade (Tim)
[ ] Add tests for gflag (Rajat)
[ ] Get consensus on syntax and keywords (Sushant)
[x] Implementation of timed lock has been delayed.
[ ] Determine what type we want for the locked timestamp
[ ] Use the same update approach as updateCatalog version, if possible
[ ] Include the NOPROFILE commands
[ ] Finalize dropping semantics
[ ] Update the description of this diff with the final syntax / table definitions

Test Plan:
```
ybd --java-test TestYsqlUpgrade
ybd --java-test TestYbRoleProfile
ybd --java-test TestPgRegressProfile
```

Reviewers: smishra, skumar, jason, telgersma

Subscribers: neil, yql, jason

Differential Revision: https://phabricator.dev.yugabyte.com/D21043