refactor: major change to YbDropProfile
Change to use DropProfileStmt that calls custom YbDropProfile.  This is
similar to how DROP TABLESPACE code works.  It is in-progress work, so
tests likely fail due to some message changes and possible intermediate
code.  Example message changes are improvements to show profile name
instead of oid.  This came for free when restructuring and following
DropTableSpace.